fix: getCompletionRate 整数除法 → 浮点除法
测试驱动修正 #1：T12-3 暴露 int/int=0 的 bug，添加 (double) 强制转换

+++ src/main/java/com/studyflow/service/impl/TaskServiceImpl.java
@@ -75,6 +75,6 @@ public class TaskServiceImpl implements TaskService {
         if (total == 0) {
             return 0.0;
         }
-        return getCompletedCount() / total * 100;
+        return (double) getCompletedCount() / total * 100;
     }
 }
